requirements.txt
pandas
requests
faker
python-dotenv
SQLAlchemy
pymysql
cryptography
